(Refact) seperate useCompetitionTimer hook from Competition.tsx
